fixed syntax errors

+++ src/main/java/game/Skill.java
@@ -2,8 +2,8 @@ package src.main.java.game;
 
 public abstract class Skill {
   protected String skillName;
-  protected float critProbability;
-  protected float failProbability;
+  protected double critProbability;
+  protected double failProbability;
   protected int damagePoints;
   
   public abstract void use();
